Urban Planning
A city always contains more than any inhabitant can know, and a great city always makes the unknown and the possible spurs to the imagination. In the planning and designing of new communities, housing projects, and urban renewal, the planners both private and public, need to give explicit consideration to the kind of world that is being created for the children who will be growing up in these settings. Particular attention should be given to the opportunities which the environment presents or precludes for the involvement of children both older and younger than themselves.